Read rules from Audit kernel
